control-tui/requirements.txt
textual==0.63.6
httpx==0.27.2
pyserial==3.5
zeroconf==0.136.2
